requirements.txt
subprocess
importlib
datamodel-code-generator
pydantic
google-genai
python-dotenv
openai